requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.1.0
openai>=1.10.0,<2.0.0
colorama==0.4.6
httpx==0.25.2